refactor(core): move chunk domain types to module-core + typealias re-export
Moves 4 domain types from module-synchronizer to module-core/.../core/
model/chunk:
- BasicRecord (from storage/BasicChunkFileReader.kt)
- OcidMapping (from storage/OcidMappingFileReader.kt)
- CalculatedEquipmentItem (from domain/CalculatedEquipmentItem.kt)
- GroupedEquipmentResult (from domain/CalculatedEquipmentItem.kt)

Synchronizer re-exports via typealias in domain/ChunkDomainReexport.kt
so existing consumer imports continue to work.

Updated 8 synchronizer files' imports to point at the new path.

This is a precondition for ChunkFileReaderPort (issue #1217, Task 3).
