+++ module-core/src/main/kotlin/maple/expectation/core/model/chunk/GroupedEquipmentResult.kt
@@ -1,0 +1,9 @@
+package maple.expectation.core.model.chunk
+
+data class GroupedEquipmentResult(
+    val readKey: String,
+    val ocid: String,
+    val presetNo: Int,
+    val userIgn: String? = null,
+    val items: List<CalculatedEquipmentItem>,
+)
