+++ module-core/src/main/kotlin/maple/expectation/core/model/chunk/OcidMapping.kt
@@ -1,0 +1,6 @@
+package maple.expectation.core.model.chunk
+
+data class OcidMapping(
+    val userIgn: String,
+    val ocid: String,
+)
