+++ module-synchronizer/src/main/kotlin/maple/synchronizer/domain/ChunkDomainReexport.kt
@@ -1,0 +1,6 @@
+package maple.synchronizer.domain
+
+typealias BasicRecord = maple.expectation.core.model.chunk.BasicRecord
+typealias OcidMapping = maple.expectation.core.model.chunk.OcidMapping
+typealias CalculatedEquipmentItem = maple.expectation.core.model.chunk.CalculatedEquipmentItem
+typealias GroupedEquipmentResult = maple.expectation.core.model.chunk.GroupedEquipmentResult
